profile-service: Combine schema and tool handler
